ranger/pyproject.toml
[project]
name = "ranger"
version = "0.1.0"
description = "Park assistant"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.3",
    "smolagents>=1.16.1",
    "fire",
    "python-dotenv>=1.1.0",
    "googlemaps>=4.10.0",
    "rich",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["cli", "router", "weather"]

[tool.uv]
package = true

[project.scripts]
ranger = "cli:main"
